+++ Tests/UITests/TestApp/HealthKitTestsView/HealthKitTestsView.swift
@@ -14,22 +14,6 @@ import SwiftUI
 
 
 struct HealthKitTestsView: View {
-    private enum HealthDataDeletionSelector {
-        /// Delete **everything**.
-        case allData
-        /// Delete all data saved to HealthKit by the TestApp.
-        case ourData
-        
-        var queryPredicate: NSPredicate? {
-            switch self {
-            case .allData:
-                nil
-            case .ourData:
-                HKQuery.predicateForObjects(from: HKSource.default())
-            }
-        }
-    }
-    
     @Environment(HealthKit.self) var healthKit
     @Environment(FakeHealthStore.self) var fakeHealthStore
     
@@ -60,6 +44,9 @@ struct HealthKitTestsView: View {
                 }
             }
             Section {
+                NavigationLink("Collect Samples") {
+                    CollectSamplesTestView()
+                }
                 NavigationLink("Samples Query") {
                     SamplesQueryView()
                 }
@@ -82,74 +69,16 @@ struct HealthKitTestsView: View {
                     SourceFilteredQueryView()
                 }
             }
-            Section("Collected Samples Since App Launch") {
-                let samplesBySampleType = fakeHealthStore.samples.grouped(by: \.sampleType.identifier)
-                ForEach(samplesBySampleType.sorted(using: KeyPathComparator(\.key)), id: \.key) { (entry: (String, [HKSample])) in
-                    let (sampleTypeIdentifier, samples) = entry
-                    HStack {
-                        Text(sampleTypeIdentifier)
-                            .accessibilityLabel("")
-                        Spacer()
-                        Text(String(samples.count))
-                            .accessibilityLabel("")
-                    }
-                    .accessibilityElement(children: .combine)
-                    .accessibilityLabel("\(sampleTypeIdentifier), \(samples.count)" as String)
-                }
-            }.accessibilityIdentifier("CollectedSamples")
-            if !FakeHealthStore.collectedSamplesOnly {
-                Section("Background Persistance Log") {
-                    ForEach(fakeHealthStore.backgroundPersistance) { logEntry in
-                        makeRow(for: logEntry)
-                    }
-                }
-            }
         }
         .viewStateAlert(state: $viewState)
         .task {
             await checkInitialSamplesAuthStatus()
         }
         .toolbar {
-            ToolbarItem(placement: .topBarTrailing) {
-                addTestDataToolbarItem
+            ToolbarItem(placement: .primaryAction) {
+                ActionsMenu(viewState: $viewState)
             }
         }
-    }
-    
-    
-    private var addTestDataToolbarItem: some View {
-        let testData: [TestDataDefinition] = [
-            .init(sampleType: .heartRate, samples: [
-                .init(date: .now, value: 87, unit: .count() / .minute())
-            ]),
-            .init(sampleType: .activeEnergyBurned, samples: [
-                .init(date: .now, value: 71.2, unit: .largeCalorie())
-            ]),
-            .init(sampleType: .stepCount, samples: [
-                .init(date: .now, value: 152, unit: .count())
-            ]),
-            .init(sampleType: .height, samples: [
-                .init(date: .now, value: 187, unit: .meterUnit(with: .centi))
-            ])
-        ]
-        return Menu {
-            AsyncButton("Delete Test Data from HealthKit", role: .destructive, state: $viewState) {
-                try await deleteTestData(.ourData)
-            }
-            AsyncButton("Delete all Data from HealthKit", role: .destructive, state: $viewState) {
-                try await deleteTestData(.allData)
-            }
-            Divider()
-            ForEach(testData, id: \.self) { entry in
-                AsyncButton("Add Sample: \(entry.sampleType.displayTitle)", state: $viewState) {
-                    try await addTestData([entry])
-                }
-            }
-        } label: {
-            Image(systemName: "plus")
-                .accessibilityLabel("Add")
-        }
-        .accessibilityIdentifier("Add")
     }
     
     
@@ -163,87 +92,6 @@ struct HealthKitTestsView: View {
             healthKit.didAskForAuthorization(toWrite: type)
         }
         allInitialSampleTypesAreAuthorized = readFullyAuthd && writeFullyAuthd
-    }
-    
-    
-    @ViewBuilder
-    private func makeRow(for logEntry: BackgroundDataCollectionLogEntry) -> some View {
-        VStack(alignment: .leading) {
-            HStack {
-                Text(logEntry.displayTitle.uppercased())
-                    .font(.system(size: 11, weight: .semibold))
-                    .foregroundStyle(.secondary)
-                Spacer()
-                Text(logEntry.id.uuidString).font(.caption2.monospaced())
-            }
-            switch logEntry {
-            case let .added(id: _, type, date, quantity):
-                Text(type)
-                    .font(.caption.monospaced())
-                HStack {
-                    if date.lowerBound == date.upperBound {
-                        Text(date.lowerBound.formatted(date: .abbreviated, time: .shortened))
-                    } else {
-                        let start = date.lowerBound.formatted(date: .abbreviated, time: .shortened)
-                        let endUsesOnlyTime = date.upperBound.timeIntervalSince(date.lowerBound) < 60 * 60 * 24
-                        let end = date.upperBound.formatted(date: endUsesOnlyTime ? .omitted : .abbreviated, time: .shortened)
-                        Text("\(start) – \(end)")
-                    }
-                    if let quantity {
-                        Spacer()
-                        Text(quantity)
-                    }
-                }
-            case .removed:
-                EmptyView()
-            }
-        }
-    }
-    
-    
-    // MARK: Test Data Handling
-    
-    private func addTestData(_ definitions: [TestDataDefinition]) async throws {
-        let samples: [HKQuantitySample] = definitions.flatMap { definition in
-            definition.samples.map { sampleInput in
-                let date = sampleInput.date
-                return HKQuantitySample(
-                    type: definition.sampleType.hkSampleType,
-                    quantity: HKQuantity(unit: sampleInput.unit, doubleValue: sampleInput.value),
-                    start: date,
-                    end: date.addingTimeInterval(sampleInput.duration)
-                )
-            }
-        }
-        try await healthKit.askForAuthorization(for: .init(write: samples.mapIntoSet(\.sampleType)))
-        for sample in samples {
-            // NOTE: for some reason, this works but calling the overload that takes an array doesn't...
-            try await healthKit.healthStore.save(sample)
-        }
-    }
-    
-    
-    private func deleteTestData(_ selector: HealthDataDeletionSelector) async throws {
-        for sampleType in healthKit.dataAccessRequirements.write {
-            let descriptor = HKSampleQueryDescriptor(
-                predicates: [
-                    HKSamplePredicate<HKSample>.sample(
-                        type: sampleType,
-                        predicate: selector.queryPredicate
-                    )
-                ],
-                sortDescriptors: []
-            )
-            do {
-                let samples = (try? await descriptor.result(for: healthKit.healthStore)) ?? []
-                if !samples.isEmpty {
-                    try await healthKit.healthStore.delete(samples)
-                }
-            } catch {
-                print("Failed to delete \(sampleType): \(error)")
-                throw error
-            }
-        }
     }
 }
